big refactor. move loading snippets from extension.ts have no idea what I've done. tests are green fix when npm dep set in defaults
