+++ .vscode/launch.json
@@ -12,6 +12,25 @@
             ],
             "request": "launch",
             "type": "extensionHost"
-        }
+        },
+        // NEVER USE THIS!
+        // {
+        //     "name": "Launch Tests",
+        //     "type": "extensionHost",
+        //     "request": "launch",
+        //     "runtimeExecutable": "${execPath}",
+        //     "args": [
+        //         "--disable-extensions",
+        //         "--extensionDevelopmentPath=${workspaceRoot}/out",
+        //         "--extensionTestsPath=${workspaceRoot}/testsOut"
+        //     ],
+        //     "env": {
+        //         "NODE_ENV": "test"
+        //     },
+        //     "sourceMaps": true,
+        //     "outFiles": [
+        //         "${workspaceRoot}/out/*.js"
+        //     ],
+        // }
     ]
 }
